Split API in 2 libraries: libnorns and libnornsctl
User-level functions are provided by libnorns.so,
which communicates with the daemon through the
norns_api_global_socket. This socket should allow
anyone to read and write from/to it.

Admin-level functions are provided by
libnornsctl.so, which communicates with urd
through the norns_api_control_socket. This socket
should allow only selected processes to read and
write from/to it.