Rename norns_status() to norns_error()
We need to provide a nornsctl_status() function to check the status of
administrative data transfer tasks, but we already have
a nornsctl_status() function in the API that checks the global status of
the service. Thus, we decided to keep the current nornsctl_status() and
rename norns_status() to norns_error() and provide a nornsctl_error()
for simmetry.